feat[build]: Startup -> organization of services
