+++ Program.cs
@@ -1,57 +1,19 @@
-using Microsoft.EntityFrameworkCore;
-using ECommerce_ASP_NET_API.Context;
-using ECommerce_ASP_NET_API.Modules.Customer;
-using ECommerce_ASP_NET_API.Modules.Customer.Contracts;
-using ECommerce_ASP_NET_API.Modules.Category;
-using ECommerce_ASP_NET_API.Modules.Category.Contracts;
+using ECommerce_ASP_NET_API.Startup;
 
 var builder = WebApplication.CreateBuilder(args);
 
-#region @Service
-builder.Services.AddControllers();
-builder.Services.AddEndpointsApiExplorer();
-builder.Services.AddSwaggerGen();
-#endregion
+ServiceProviders.Controller(builder);
 
+ServiceProviders.Swagger(builder);
 
-#region @Database
-builder.Services.AddDbContext<DatabaseContext>(options =>
-{
-    var host = builder.Configuration.GetConnectionString("DefaultConnection");
+ServiceProviders.Database(builder);
 
-    options.UseMySql(
-        connectionString: host,
-        serverVersion: ServerVersion.AutoDetect(host));
-});
-
-builder.Services.AddAutoMapper(AppDomain.CurrentDomain.GetAssemblies());
-#endregion
-
-
-#region @Injectable
-builder.Services.AddScoped<ICategoryRepository, CategoryRepository>();
-builder.Services.AddScoped<ICategoryService, CategoryService>();
-builder.Services.AddScoped<ICustomerRepository, CustomerRepository>();
-builder.Services.AddScoped<ICustomerService, CustomerService>();
-#endregion
-
+ServiceProviders.Injectable(builder);
 
 var app = builder.Build();
 
+ServiceProviders.Development(app);
 
-#region @Pipeline
-if (app.Environment.IsDevelopment())
-{
-    app.UseSwagger();
-    app.UseSwaggerUI();
-}
-
-app.UseHttpsRedirection();
-
-app.UseAuthorization();
-
-app.MapControllers();
-#endregion
-
+ServiceProviders.Middleware(app);
 
 app.Run();
