+++ Startup/ProviderSwagger.cs
@@ -1,0 +1,9 @@
+namespace ECommerce_ASP_NET_API.Startup;
+
+public static partial class ServiceProviders
+{
+    public static void Swagger(WebApplicationBuilder builder)
+    {
+        builder.Services.AddSwaggerGen();
+    }
+}
